+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 
 ### Added
+* Support Tectonic V2 CLI in run configuration
+* Add basic support for multiple inputs in Tectonic.toml
+* Improve performance of file set cache used by inspections
 * Support label references to user defined listings environment
 * Add option to disable automatic compilation in power save mode
 * Convert automatic compilation settings to a combobox